Consolidate exceptions to a package
