+++ tests/test_entities_config.py
@@ -4,8 +4,8 @@ from lib.AccountsConfig import AccountsConfig
 from lib.PartiesConfig import PartiesConfig
 from lib.PartyAddressesConfig import PartyAddressesConfig
 
-from lib.MissingSourceLocationError import MissingSourceLocationError
-from lib.MissingSchemaError import MissingSchemaError
+from exceptions.MissingSourceLocationError import MissingSourceLocationError
+from exceptions.MissingSchemaError import MissingSchemaError
 
 
 def test_accounts_config_throws_when_source_location_is_missing():
